changing the logic of details screen from A to Z -add new path to API to fetch book details -add new path to API to fetch book ratings -fetch the book the second we open it from home screen -add a lot of models to details screen *still the author name need to be added to API path and fetching it

+++ lib/Features/details/data/models/ratings_response.dart
@@ -15,7 +15,7 @@ class RatingsResponse {
 
 @JsonSerializable()
 class Rating {
-  final int? average;
+  final double? average;
   final int? count;
 
   Rating({this.average, this.count});
